requirements.txt
flask==3.0.0
anthropic>=0.86.0
ebooklib==0.18
python-dotenv==1.0.0
werkzeug==3.0.1
beautifulsoup4==4.12.2
lxml==6.0.2
Pillow==12.1.0
